Package structure: moved sensors.
Part of #963.
