+++ src/main/java/de/dennisguse/opentracks/sensors/PressureSensorUtils.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks.services.sensors;
+package de.dennisguse.opentracks.sensors;
 
 import android.hardware.SensorManager;
